lambda/notify-to-app/requirements.txt
beautifulsoup4
boto3
cloudscraper
urllib3 <2.0.0,>=1.26.0
strands-agents
strands-agents-builder